+++ packages/adapter/src/request-building/authentication.test.ts
