+++ packages/adapter/src/request-building/authentication.ts
