+++ packages/adapter/src/response-processing/casting.test.ts
