+++ packages/adapter/src/response-processing/encoding.test.ts
