+++ packages/adapter/src/response-processing/extraction.test.ts
