+++ packages/adapter/src/response-processing/index.test.ts
